pyproject.toml
[tool.poetry]
name = "text2sql"
version = "0.1.0"
description = ""
authors = ["Haohanblue <haohanblue@foxmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymysql = "^1.1.1"
zhipuai = "^2.1.5.20250106"
python-dotenv = "^1.1.0"
requests = "^2.32.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
